[s-ramp-ide] apply corrections
