+++ s-ramp-ide/eclipse/repository-browser.test/META-INF/MANIFEST.MF
