+++ s-ramp-ide/eclipse/repository-browser/META-INF/MANIFEST.MF
@@ -1,0 +1,48 @@
+Manifest-Version: 1.0
+Bundle-ManifestVersion: 2
+Bundle-Name: srampRepositoryBrowser
+Bundle-SymbolicName: srampRepositoryBrowser;singleton:=true
+Bundle-Version: 0.7.0.Final
+Bundle-Activator: org.overlord.sramp.srampRepositoryBrowser.Activator
+Bundle-RequiredExecutionEnvironment: JavaSE-1.7
+Require-Bundle: org.eclipse.ui,
+ org.eclipse.core.runtime,
+ org.eclipse.core.resources
+Bundle-ActivationPolicy: lazy
+Bundle-ClassPath: .,
+ /lib/activation.jar,
+ /lib/apache-mime4j.jar,
+ /lib/commons-beanutils-core.jar,
+ /lib/commons-codec.jar,
+ /lib/commons-collections.jar,
+ /lib/commons-configuration.jar,
+ /lib/commons-digester.jar,
+ /lib/commons-io.jar,
+ /lib/commons-lang.jar,
+ /lib/commons-logging.jar,
+ /lib/httpclient.jar,
+ /lib/httpcore.jar,
+ /lib/javassist.jar,
+ /lib/jaxb-impl.jar,
+ /lib/jaxrs-api.jar,
+ /lib/jboss-logging.jar,
+ /lib/jcip-annotations.jar,
+ /lib/joda-time.jar,
+ /lib/jsr250-api.jar,
+ /lib/log4j.jar,
+ /lib/mail.jar,
+ /lib/org.apache.felix.scr.annotations.jar,
+ /lib/out,
+ /lib/overlord-commons-auth.jar,
+ /lib/overlord-commons-config.jar,
+ /lib/resteasy-atom-provider.jar,
+ /lib/resteasy-jaxb-provider.jar,
+ /lib/resteasy-jaxrs.jar,
+ /lib/resteasy-multipart-provider.jar,
+ /lib/scannotation.jar,
+ /lib/slf4j-api.jar,
+ /lib/slf4j-log4j12.jar,
+ /lib/s-ramp-api.jar,
+ /lib/s-ramp-atom.jar,
+ /lib/s-ramp-client.jar,
+ /lib/s-ramp-common.jar
