+++ s-ramp-ide/eclipse/repository-browser/build.properties
@@ -1,0 +1,9 @@
+source.. = src/main/java/
+output.. = target/classes
+src.includes = pom.xml
+bin.includes = plugin.xml,\
+               META-INF/,\
+               .,\
+               lib/,\
+               icons/,\
+              contexts.xml
