+++ s-ramp-ide/eclipse/repository-browser/icons/sample.gif
