+++ s-ramp-ide/eclipse/repository-browser/plugin.xml
@@ -1,0 +1,136 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<?eclipse version="3.4"?>
+<plugin>
+
+   <extension
+         point="org.eclipse.ui.views">
+      <category
+            name="S-RAMP"
+            id="sramp">
+      </category>
+      <view
+            category="sramp"
+            class="org.overlord.sramp.srampRepositoryBrowser.commands.RepositoryBrowser"
+            icon="icons/sample.gif"
+            id="org.overlord.sramp.srampRepositoryBrowser.views.RepositoryBrowser"
+            name="Repository Browser">
+      </view>
+   </extension>
+   <extension
+         point="org.eclipse.ui.perspectiveExtensions">
+      <perspectiveExtension
+            targetID="org.eclipse.jdt.ui.JavaPerspective">
+         <view
+               ratio="0.5"
+               relative="org.eclipse.ui.views.ProblemView"
+               relationship="right"
+               id="org.overlord.sramp.srampRepositoryBrowser.views.RepositoryBrowser">
+         </view>
+      </perspectiveExtension>
+   </extension>
+   <extension
+         point="org.eclipse.help.contexts">
+      <contexts
+            file="contexts.xml">
+      </contexts>
+   </extension>
+   <extension
+         point="org.eclipse.ui.commands">
+      <command
+            defaultHandler="org.overlord.sramp.srampRepositoryBrowser.commands.ConnectToServerComand"
+            id="srampRepositoryBrowser.ConnectToServer"
+            name="ConnectToServer">
+      </command>
+      <command
+            defaultHandler="org.overlord.sramp.srampRepositoryBrowser.commands.ImportArtifactComand"
+            id="srampRepositoryBrowser.ImportToSramp"
+            name="ImportToSramp">
+      </command>
+   </extension>
+   
+      
+   <extension
+         point="org.eclipse.ui.menus">
+      <menuContribution
+            allPopups="false"
+            locationURI="toolbar:org.overlord.sramp.srampRepositoryBrowser.views.RepositoryBrowser">
+         <toolbar
+               id="srampRepositoryBrowser.toolbar4">
+            <command
+                  commandId="srampRepositoryBrowser.ConnectToServer"
+                  style="push">
+            </command>
+         </toolbar>
+         <command
+               commandId="srampRepositoryBrowser.ConnectToServer"
+               label="Connect To Server"
+               style="push">
+         
+		
+           </command>
+      </menuContribution>
+      <menuContribution
+            allPopups="false"
+            locationURI="popup:org.eclipse.jdt.ui.PackageExplorer">
+         <command
+               commandId="srampRepositoryBrowser.ImportToSramp"
+               label="Import to S-RAMP repository"
+               style="push">
+             <visibleWhen>
+				           <with variable="activeMenuSelection">
+				            <iterate
+				                 ifEmpty="false">
+				             <adapt type="org.eclipse.core.resources.IResource">
+				             <instanceof
+                           value="org.eclipse.core.resources.IFile"/>
+				             </adapt>
+				            </iterate>
+				           </with>
+         			</visibleWhen> 
+               
+         </command>
+      </menuContribution>
+       <menuContribution
+            allPopups="false"
+            locationURI="popup:org.eclipse.ui.navigator.ProjectExplorer#PopupMenu">
+         <command
+               commandId="srampRepositoryBrowser.ImportToSramp"
+               label="Import to S-RAMP repository"
+               style="push">
+               <visibleWhen>
+				           <with variable="activeMenuSelection">
+				            <iterate
+				                 ifEmpty="false">
+				             <adapt type="org.eclipse.core.resources.IResource">
+				             <instanceof
+                           value="org.eclipse.core.resources.IFile"/>
+				             </adapt>
+				            </iterate>
+				           </with>
+         			</visibleWhen>
+         </command>
+      </menuContribution>
+      
+       <menuContribution
+            allPopups="false"
+            locationURI="popup:org.eclipse.ui.views.ResourceNavigator?after=additions">
+         <command
+               commandId="srampRepositoryBrowser.ImportToSramp"
+               label="Import to S-RAMP repository"
+               style="push">
+                <visibleWhen>
+				           <with variable="activeMenuSelection">
+				            <iterate
+				                 ifEmpty="false">
+				             <adapt type="org.eclipse.core.resources.IResource">
+				             <instanceof
+                           value="org.eclipse.core.resources.IFile"/>
+				             </adapt>
+				            </iterate>
+				           </with>
+         			</visibleWhen>
+         </command>
+      </menuContribution>
+   </extension>
+   
+</plugin>
