--- s-ramp-ide/srampRepositoryBrowser/build.properties
@@ -1,9 +1,0 @@
-source.. = src/main/java
-output.. = target/classes
-src.includes = pom.xml
-bin.includes = plugin.xml,\
-               META-INF/,\
-               .,\
-               lib/,\
-               icons/,\
-              contexts.xml
