--- s-ramp-ide/srampRepositoryBrowser/src/main/java/cz/muni/fi/srampRepositoryBrowser/manager/ServiceFailureException.java
@@ -1,23 +1,0 @@
-package cz.muni.fi.srampRepositoryBrowser.manager;
-
-/**
- * service failture exception
- * 
- * @author Jan Bouska
- * 
- */
-public class ServiceFailureException extends RuntimeException {
-
-	public ServiceFailureException(String msg) {
-		super(msg);
-	}
-
-	public ServiceFailureException(Throwable cause) {
-		super(cause);
-	}
-
-	public ServiceFailureException(String message, Throwable cause) {
-		super(message, cause);
-	}
-
-}
